+++ lib/html_pipeline/node_filter/syntax_highlight_filter.rb
@@ -1,0 +1,50 @@
+# frozen_string_literal: true
+
+HTMLPipeline.require_dependency("rouge", "SyntaxHighlightFilter")
+
+class HTMLPipeline
+  class NodeFilter
+    # HTML Filter that syntax highlights text inside code blocks.
+    #
+    # Context options:
+    #
+    #   :highlight => String represents the language to pick lexer. Defaults to empty string.
+    #   :scope => String represents the class attribute adds to pre element after.
+    #             Defaults to "highlight highlight-css" if highlights a css code block.
+    #
+    # This filter does not write any additional information to the context hash.
+    class SyntaxHighlightFilter < NodeFilter
+      def initialize(doc, context: {}, result: {})
+        super(doc, context: context, result: result)
+        @formatter = Rouge::Formatters::HTML.new
+      end
+
+      def call
+        doc.search("pre").each do |node|
+          default = context[:highlight]&.to_s
+          next unless (lang = node["lang"] || default)
+          next unless (lexer = lexer_for(lang))
+
+          text = node.inner_text
+          html = highlight_with_timeout_handling(text, lexer)
+          next if html.nil?
+
+          node.inner_html = html
+          scope = context.fetch(:scope, "highlight")
+          node["class"] = "#{scope} #{scope}-#{lang}"
+        end
+        doc
+      end
+
+      def highlight_with_timeout_handling(text, lexer)
+        Rouge.highlight(text, lexer, @formatter)
+      rescue Timeout::Error => _e
+        nil
+      end
+
+      def lexer_for(lang)
+        Rouge::Lexer.find(lang)
+      end
+    end
+  end
+end
