--- lib/html_pipeline/plain_text_input_filter.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-HTMLPipeline.require_dependency("escape_utils", "PlainTextInputFilter")
-
-class HTMLPipeline
-  # Simple filter for plain text input. HTML escapes the text input and wraps it
-  # in a div.
-  class PlainTextInputFilter < TextFilter
-    def call
-      "<div>#{EscapeUtils.escape_html(@text, false)}</div>"
-    end
-  end
-end
