+++ lib/html_pipeline/text_filter.rb
@@ -9,7 +9,17 @@ class HTMLPipeline
 
       # Ensure that this is always a string
       @text = text.respond_to?(:to_str) ? text.to_str : text.to_s
-      super(nil, context: context, result: result)
+      super(context: context, result: result)
+    end
+
+    # Like call but guarantees that a string of HTML markup is returned.
+    def self.process(input, context: {})
+      output = call(input, context: context)
+      if output.respond_to?(:to_html)
+        output.to_html
+      else
+        output.to_s
+      end
     end
   end
 end
