+++ test/html_pipeline/filter/node_filter/image_max_width_filter_test.rb
@@ -5,7 +5,7 @@ require "test_helper"
 class HTMLPipeline
   class ImageMaxWidthFilterTest < Minitest::Test
     def setup
-      @filter = HTMLPipeline::ImageMaxWidthFilter
+      @filter = HTMLPipeline::NodeFilter::ImageMaxWidthFilter
     end
 
     def test_rewrites_image_style_tags
