+++ test/html_pipeline/filter/text_filter/image_filter_test.rb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-ImageFilter = HTMLPipeline::ImageFilter
+ImageFilter = HTMLPipeline::TextFilter::ImageFilter
 
 class HTMLPipeline
   class ImageFilterTest < Minitest::Test
@@ -12,32 +12,32 @@ class HTMLPipeline
 
     def test_jpg
       assert_equal(%(<img src="http://example.com/test.jpg" alt=""/>),
-        @filter.to_html(%(http://example.com/test.jpg)))
+        @filter.call(%(http://example.com/test.jpg)))
     end
 
     def test_jpeg
       assert_equal(%(<img src="http://example.com/test.jpeg" alt=""/>),
-        @filter.to_html(%(http://example.com/test.jpeg)))
+        @filter.call(%(http://example.com/test.jpeg)))
     end
 
     def test_bmp
       assert_equal(%(<img src="http://example.com/test.bmp" alt=""/>),
-        @filter.to_html(%(http://example.com/test.bmp)))
+        @filter.call(%(http://example.com/test.bmp)))
     end
 
     def test_gif
       assert_equal(%(<img src="http://example.com/test.gif" alt=""/>),
-        @filter.to_html(%(http://example.com/test.gif)))
+        @filter.call(%(http://example.com/test.gif)))
     end
 
     def test_png
       assert_equal(%(<img src="http://example.com/test.png" alt=""/>),
-        @filter.to_html(%(http://example.com/test.png)))
+        @filter.call(%(http://example.com/test.png)))
     end
 
     def test_https_url
       assert_equal(%(<img src="https://example.com/test.png" alt=""/>),
-        @filter.to_html(%(https://example.com/test.png)))
+        @filter.call(%(https://example.com/test.png)))
     end
   end
 end
